Move Card, Message, User to entities package
