+++ app/src/main/java/com/shapeapp/shape/backendcalls/BackendApi.kt
@@ -1,6 +1,6 @@
 package com.shapeapp.shape.backendcalls
 
-import com.shapeapp.shape.data.Card
+import com.shapeapp.shape.data.entities.Card
 import retrofit2.Call
 import retrofit2.http.GET
