+++ app/src/main/java/com/shapeapp/shape/recyclerviewinterfaces/RecyclerViewMessageClickListener.kt
@@ -2,7 +2,7 @@ package com.shapeapp.shape.recyclerviewinterfaces
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
-import com.shapeapp.shape.data.Message
+import com.shapeapp.shape.data.entities.Message
 
 /**
  * Interface definition for a callback to be invoked when a message in [RecyclerView] is clicked.
