+++ app/src/main/java/com/shapeapp/shape/repositories/MessagesRepository.kt
@@ -2,7 +2,7 @@ package com.shapeapp.shape.repositories
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import com.shapeapp.shape.data.Message
+import com.shapeapp.shape.data.entities.Message
 import com.shapeapp.shape.mockupsmakers.MessageMockups
 
 /**
