+++ app/src/main/java/com/shapeapp/shape/ui/messages/MessagesFragment.kt
@@ -11,7 +11,7 @@ import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.shapeapp.shape.R
-import com.shapeapp.shape.data.Message
+import com.shapeapp.shape.data.entities.Message
 import com.shapeapp.shape.mockupsmakers.MessageMockups
 import com.shapeapp.shape.mockupsmakers.UsersMockups
 import com.shapeapp.shape.recyclerviewadapters.MessagesPreviewRecyclerViewAdapter
