+++ app/src/main/java/com/shapeapp/shape/ui/messages/MessagesViewModel.kt
@@ -2,7 +2,7 @@ package com.shapeapp.shape.ui.messages
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
-import com.shapeapp.shape.data.Message
+import com.shapeapp.shape.data.entities.Message
 import com.shapeapp.shape.repositories.MessagesRepository
 
 /**
