+++ app/src/main/java/com/shapeapp/shape/ui/publicpackage/PublicViewModel.kt
@@ -2,7 +2,7 @@ package com.shapeapp.shape.ui.publicpackage
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
-import com.shapeapp.shape.data.Card
+import com.shapeapp.shape.data.entities.Card
 import com.shapeapp.shape.repositories.CardsRepository
 
 /**
