+++ app/src/main/java/com/shapeapp/shape/ui/receivedimage/ReceivedImageFragment.kt
@@ -10,7 +10,7 @@ import androidx.lifecycle.ViewModelProviders
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
 import com.shapeapp.shape.R
-import com.shapeapp.shape.data.Card
+import com.shapeapp.shape.data.entities.Card
 import com.shapeapp.shape.gesturesdetection.OnFourWaysSwipeListener
 import kotlinx.android.synthetic.main.fragment_received_image.*
