+++ app/src/main/java/com/shapeapp/shape/ui/receivedimage/ReceivedImageViewModel.kt
@@ -4,7 +4,7 @@ import android.net.Uri
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.shapeapp.shape.data.Card
+import com.shapeapp.shape.data.entities.Card
 
 /**
  * [ViewModel] for [ReceivedImageFragment]
