+++ app/src/main/res/navigation/navigation_graph.xml
@@ -32,7 +32,7 @@
         tools:layout="@layout/fragment_profile" >
         <argument
             android:name="user"
-            app:argType="com.shapeapp.shape.data.User" />
+            app:argType="com.shapeapp.shape.data.entities.User" />
     </fragment>
     <fragment
         android:id="@+id/messagesFragment"
@@ -66,7 +66,7 @@
         tools:layout="@layout/fragment_received_image">
         <argument
             android:name="card"
-            app:argType="com.shapeapp.shape.data.Card" />
+            app:argType="com.shapeapp.shape.data.entities.Card" />
         <action
             android:id="@+id/action_receivedImageFragment_to_publicFragment"
             app:destination="@id/publicFragment"
@@ -84,7 +84,7 @@
         tools:layout="@layout/fragment_chat">
         <argument
             android:name="messages"
-            app:argType="com.shapeapp.shape.data.Message[]" />
+            app:argType="com.shapeapp.shape.data.entities.Message[]" />
     </fragment>
     <fragment
         android:id="@+id/cameraFragment"
